refacto FolderList, MediaLibrary use of FolderList
