+++ packages/core/upload/admin/src/translations/en.json
@@ -33,6 +33,7 @@
   "input.placeholder.icon": "Drop the asset in this zone",
   "input.url.description": "Separate your URL links by a carriage return.",
   "input.url.label": "URL",
+  "list.assets.title": "Assets",
   "list.asset.at.finished": "The assets have finished loading.",
   "list.asset.load": "Loading the asset list.",
   "list.assets-empty.search": "No result found",
@@ -46,6 +47,8 @@
   "list.assets.preview-asset": "Preview for the video at path {path}",
   "list.assets.selected": "{number, plural, =0 {No asset} one {1 asset} other {# assets}} ready to upload",
   "list.assets.type-not-allowed": "This type of file is not allowed.",
+  "list.folder.edit": "Edit folder",
+  "list.folder.subtitle": "{folderCount} folder, {filesCount} assets",
   "mediaLibraryInput.actions.nextSlide": "Next slide",
   "mediaLibraryInput.actions.previousSlide": "Previous slide",
   "mediaLibraryInput.placeholder": "Click to add an asset or drag and drop one in this area",
